requirements.txt
asgiref==3.8.1
boto3==1.34.91
botocore==1.34.91
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.8
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.7
jmespath==1.0.1
Markdown==3.6
mysqlclient==2.2.7
# mysqlclient
numpy==1.26.4
openpyxl==3.1.4
pandas==2.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.8.2
django-cors-headers==4.7.0
